requirements.txt
streamlit>=1.40.0
python-dotenv>=1.0.0
requests>=2.28.0
markdown>=3.5.0
bleach>=6.0.0
